Fix to enable execution of builds on master computer
This checkin enables gearman to execute builds on the master node.  The
Master node is a special case because it does not inherit from the Slave Node object in
Jenkins, instead it's a Computer object so whenever you do Jenkins.getNodes() the master
is not in the list of Nodes. Also the master name is "", however when you want to
run a build against the master you need to pass in a "master" label to scheduleBuild
method.

Change-Id: I65c21e7cf7f2e244c94638f8858ab0fa5fc8acb9