abort running jenkins builds
This checkin will allow a gearman client to request a job to abort a
build that Jenkins is currently running.  The client request needs to
send in the build id to request an abort.  This only aborts builds
that have already made it to jenkins.  It will NOT abort builds that are
still on the gearman queue. Canceling jobs from the gearman queue is a TODO

Added pydev nature due to the example python script added in last commit.
Also setup auto formatting to help remove trailing white spaces.

add additional documentation

set the plugin version starting at 0.0.1

Change-Id: I533934723ee4f4814ebd68c8a2f9fe32a6aef718