give each thread an id
This checkin adds an id to each gearman executor.  Not sure what we
wanted to use for the UUID at this time so I've made it a UUID.  Also
added a constructor that doesn't require user to specify a  name for the
thread.  By default name will be set to 'anonymous' if not passed in.
Created getters and setters for object variables.

AbstractWorkerThread - auto generate a random id for each executor and
	set name to 'anonymous' if user doesn't specify.
ExecutorWorkerThread - added constructor
ManagementWorkerThread - added constructor

Change-Id: I82958d259ca951ed08cccebc369fa62c44ab532f