make gearman-plugin enable/disable state persistant across jenkins restart
This checkin will enable the gearman-plugin to be in the same state
as it was configured to before jenkins shutdown.  The gearman plugin will
now startup in ether enabled or disabled state.  Enabled means that
the plugin will automatically spawn all the threads and register all the
functions depending on the state of the slaves, projects and labels.
Disabled means that no workers are spawned and no functions are registers.
When in diabled state user can manually enable the gearman-plugin from the
jenkins configuration page.  If the plugin is enabled on startup the user can
disable it from the jenkins global configuration page, which will stop
all of the workers.

src/main/java/hudson/plugins/gearman/ComputerListenerImpl.java
    automate the spawning of executors when master comes online.
    add license info

src/main/java/hudson/plugins/gearman/GearmanPluginConfig.java
    remove setting to put gearman-plugin in disable state on startup.

src/main/java/hudson/plugins/gearman/GearmanProxy.java
    Add method to get the list of management executors
    changed thead list from Stack to ArrayList object

Change-Id: Iea20dd244b322de0bf75eff3b31a4f70ed688282