gearman function update on jenkins node label change
This checkin enables the gearman plugin to update gearman
functions on a node label change.  We use a SaveableListener to
listen for changes to nodes and projects.  The SaveableListener
is a catch all for any event that requires a save.  If any change
happens then we re-register all gearman functions in every
executor worker thread.  This listener essentially replaces the
ProjectListener and some of the ComputerListener events.

We needed to use the SaveableListener as a workaround
for the ComputerListener onConfigurationChange method due to
a bug in the onConfigurationChange method.  The jenkins
config doesn't get saved until after the onConfigurationChange
is called.

This seems to work ok but probably not a good solution because
the SaveableListener fires for every instance of a jenkins node.
So if there are 3 nodes in jenkins it will fire 3 time even though
there was only a change to one of the nodes. This will slow down
the system as you have more executors.

This is also not a great solution because
I'm sensing a race condition that may conflict
between SaveablListner and ComputerListener.  On a delete or
disconnect of a node the disconnect may happen during an
gearman function update which may cause a NPE during the update.

This checkin also enables updating gearman functions on
a project state change (disable/enable).

I believe the only event left that requires gearman funciton
updates are adding and removing of jenkins executors.

Change-Id: I67747d4d895a2474477d2a8ec01c3226e1d5456a