Add local GearmanWorker.
Add a package local implementation of something like the GearmanWorker from
java-gearman (based on GearmanWorkerImpl).

It is much simpler than the existing GearmanWorkerImpl and is more suited to
the way we need to use it in the Jenkins plugin.  It assumes jobs are always
changed in batches, and only changes jobs at the top of the event loop (not
when a job is running).

The worker threads are updated to only request job changes when there is an
actual difference.

WORK_STATUS events are sent every 10 seconds while a job is running.

run-fast is updated to only remove the gearman plugin from the work directory,
preserving any other plugins that may be installed.

This isn't very elegant, but is a start and broadly demonstrates what we need
the plugin to do.

Change-Id: I26df504534ec50f03c9e0ef772a709046cf88a23