Define a unique parameter for stop worker.
The stop worker identifies the job that should be stopped by the
unique ID supplied when the job is started.  However, the current
implementation gets the unique ID of the job to cancel from the unique
id of the stop job itself.  This is confusing and possibly erroneous
since it overloads the meaning of a gearman unique id.

Instead, have the stop job interpret the gearman argument as the
unique id of the job to stop.

Change-Id: Ida7a68976d3f764df504aff9afdae81265e0006f