add gearman worker to set build descriptions
This commit adds another management worker for setting a build's descriptions.
The worker will lookup the build by it's build_id (jobName:jobId)
and if it can find it then will set the description and return with COMPLETED.
If it cannot find the build or cannot set the description for any reason
it will return FAILED.

This commit also adds teh build_id to the data returned by the StartJobWorker.

Usage:

  client = gear.Client()
  client.addServer('localhost')
  client.waitForServer()
  job_name = 'set_description:JenkinsHostName'
  data = {'build_id':"pep8:2013-05-15_17-32-07",'description':"<h1>My Test Build</h1>"}
  job = gear.Job(job_name, simplejson.dumps(data))
  client.submitJob(job)

Change-Id: I4990772d591d27bbb3b4f20abfb4d988077b4995