Handle not connecting at start.
We previously assumed that we would connect to gearman immediately and be able
to send a SET_CLIENT_ID packet; that would raise an exception and stop any
further gearman processing.  Instead, don't assume we can send that packet,
and also, do send it immediately after reconnecting.

Also, lower the retry delay from 3 second to 2.  I don't think we'll mind
Jenkins being in a tighter loop because it probably won't be doing anything else.

Change-Id: Id585726e23076ad22e2935173995db2bdadf9974