always run management worker
This commit is a fix to bug 1185952.

While the jenkins host is running its master node may
be in a disabled state (master has 0 executors) however the
host should still be able to run a management worker and
communicate with a Gearman Server.  This commit makes
sure that the manaegement worker runs regardless
the master node's state.

This commit also makes sure that we are setting the management
worker's name to something.

Change-Id: I871019fadde47787cfe20bec666e24d242728f30