Fix project-node registration
Previously, a node would register the ability to run a job under
all labels associated with the project if the node matched at
least one of them.  What we really want to do is, if the node
matches any project label, register the generalized job and then
also register it for each label in the _intersection_ of project
labels and node labels.

Change-Id: Ie41af7b1a2f26a4ca2aac6c7263817d5021cca9b