Use TextParameterValue instead of String
If a value with a newline is received, Jenkins does not display the
value correctly in the parameters page.  Based on a quick reading
of similar issues elsewhere, it may also not save the value correctly
for later use by plugins such as the 'rebuild' plugin.  Switching
from string to text parameter types solves this.  However, it does
cause _all_ parameters to be treated as text, which wastes a bit of
real-estate on the parameters listing screen with tall textboxes.
We could scan the string for '\n' and choose appropriately as an
alternative.

Change-Id: I84ef198fd6ef852fc0a403e126f13b8cbb58a7b1