Fix deadlock from a WORK_FAIL event
A lock is kept on a gearman worker when a WORK_FAIL event happens.
This causes the worker therad to stall on the next attempt that
the AvailabilityMonitor attempt to get a lock on the worker.
This causes the jenkins nodes to stop working (not run builds
anymore). Unlock the worker on a WORK_FAIL event to avoid this
deadlock state.

This fixes issue https://issues.jenkins-ci.org/browse/JENKINS-28891

Change-Id: I015ce9732fd535676a832680f39e220b09df95cf